Move parseTextFilterString function to the appropriate utils folder
