+++ frontend/src/components/TextFilterQueryBuilder/utils/index.ts
